Map Working in Alert View
Got the map working.

There's a difference when using constraints and autolayout with these
maps in iOS 6. Before contraints a view new it's size the moment it was
intialized. With constriants it take longer (well past view did load)
to figure out the frame size for the object. So when you do graphical
operations on these maps you need to do them in -(void)viewDidAppear
instead of -(void)viewDidLoad.